// ui/src/generated/InvitePreview.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ParticipantSummary } from "./ParticipantSummary";
import type { ProjectId } from "./ProjectId";
import type { ShareRole } from "./ShareRole";

/**
 * "<host> invites you to <project>: Join?"
 */
export type InvitePreview = { host: ParticipantSummary, project: ProjectId, project_name: string, role: ShareRole, 
/**
 * Others in the session now (for avatars).
 */
online: Array<ParticipantSummary>, 
/**
 * This store already has a copy of the project (it will be reconciled).
 */
local_copy: boolean, };